add client#tlapse spec and force creation option to cli#tlapse

+++ lib/renchin/file_processor.rb
@@ -10,10 +10,12 @@ module Renchin
 
     # delete all files in directory
     def delete_directory(image_directory_path,expr)
-      Dir::foreach(image_directory_path) do |file|
-        File.delete(image_directory_path + '/' + file) if (/#{expr}$/ =~ file)
+      if exists?(image_directory_path)
+        Dir::foreach(image_directory_path) do |file|
+          File.delete(image_directory_path + '/' + file) if (/#{expr}$/ =~ file)
+        end
+        Dir::rmdir(image_directory_path)
       end
-      Dir::rmdir(image_directory_path)
     end
 
     def exists?(filename)
